+++ frontend/src/App.css
@@ -82,6 +82,11 @@ button:hover {
   margin-bottom: 10px;
 }
 
+.success {
+  color: green;
+  margin-bottom: 10px;
+}
+
 .ticket-form {
   display: flex;
   flex-direction: column;
