+++ frontend/src/components/Login.tsx
@@ -1,42 +1,42 @@
-import React, { useState } from 'react';
+import React, { useState } from "react";
 
 const Login: React.FC<{ onLogin: () => void }> = ({ onLogin }) => {
-    const [username, setUsername] = useState('');
-    const [password, setPassword] = useState('');
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
 
-    const handleSubmit = (e: React.FormEvent) => {
-        e.preventDefault();
-        // Add your authentication logic here
-        if (username === 'user' && password === 'password') {
-            onLogin();
-        } else {
-            alert('Invalid credentials');
-        }
-    };
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    // Add your authentication logic here
+    if (username === "user" && password === "password") {
+      onLogin();
+    } else {
+      alert("Invalid credentials");
+    }
+  };
 
-    return (
-        <div className="Login">
-            <form onSubmit={handleSubmit}>
-                <div>
-                    <label>Username:</label>
-                    <input 
-                        type="text" 
-                        value={username} 
-                        onChange={(e) => setUsername(e.target.value)} 
-                    />
-                </div>
-                <div>
-                    <label>Password:</label>
-                    <input 
-                        type="password" 
-                        value={password} 
-                        onChange={(e) => setPassword(e.target.value)} 
-                    />
-                </div>
-                <button type="submit">Login</button>
-            </form>
+  return (
+    <div className="Login">
+      <form onSubmit={handleSubmit}>
+        <div>
+          <label>Username:</label>
+          <input
+            type="text"
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+          />
         </div>
-    );
+        <div>
+          <label>Password:</label>
+          <input
+            type="password"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+          />
+        </div>
+        <button type="submit">Login</button>
+      </form>
+    </div>
+  );
 };
 
 export default Login;
