+++ frontend/src/components/Profile.tsx
@@ -1,12 +1,12 @@
-import React from 'react';
+import React from "react";
 
 const Profile: React.FC = () => {
-    return (
-        <div>
-            <h2>Profile</h2>
-            {/* Add your profile content here */}
-        </div>
-    );
+  return (
+    <div>
+      <h2>Profile</h2>
+      {/* Add your profile content here */}
+    </div>
+  );
 };
 
 export default Profile;
